Make the Problem class testable
Some standalone functions were implemented to be able to load from file
or stream. The Problem ctor is now accepting concrete data types and not
references to files and stuff.

The testing data for local search was inlined. On the other hand, tests
that rely on some data file are now using preproc macros to locate the
file.