Fix local search stack based rescheduling
The stack based rescheduling is useful to arrive at the last operations
quicker and stop early whenever the makespan is deteriorating. However,
operations must be added to the stack for rescheduling only when their
parents are scheduled themselves. This is now done in a first pass that
invalidates all the operations that are affected by the swap.